chore(config): moved llm specific configs
summary: some llm specific configurations were either
duplicated or inappropriately setup on the agent and
passed down. Moved max_tokens, context_window,
reasoning_effort, temperature, and additional_params
to the llm config itself. Deleted the duplicated
llm config classes in favor of a single class. Added
#[serde(deny_unknown_fields)] to both agent and llm
configs to warn the user of breaking config changes
and included a readme blurb and breaking-change doc.
Moved the lenient_int serde helpers to their own
file. Updated model_context_limit to u64 to match RIG.

ref: LOG-23546

+++ crates/aura/src/stream_events.rs
@@ -132,7 +132,7 @@ pub enum AuraStreamEvent {
         model: String,
         /// Context limit for this model (in tokens), if known
         #[serde(skip_serializing_if = "Option::is_none")]
-        model_context_limit: Option<u32>,
+        model_context_limit: Option<u64>,
         #[serde(flatten)]
         correlation: CorrelationContext,
     },
@@ -381,7 +381,7 @@ impl AuraStreamEvent {
     /// Contains model name and context limit for UI context window calculation.
     pub fn session_info(
         model: impl Into<String>,
-        model_context_limit: Option<u32>,
+        model_context_limit: Option<u64>,
         correlation: CorrelationContext,
     ) -> Self {
         Self::SessionInfo {
